requirements.txt
aiocache==0.11.1
aiofiles==0.8.0
aioredis==1.3.0
aiosqlite==0.17.0
async-timeout==4.0.2
asyncpg==0.25.0
certifi==2021.10.8
charset-normalizer==2.0.12
Deprecated==1.2.13
hiredis==2.0.0
httptools==0.4.0
idna==3.3
iso8601==0.1.16
marshmallow==3.15.0
multidict==5.2.0
packaging==21.3
pyparsing==3.0.7
pypika-tortoise==0.1.3
pytz==2021.3
requests==2.27.1
sanic==21.12.1
sanic-routing==0.7.2
sanic-scheduler==1.0.7
tortoise-orm==0.18.1
typing-extensions==4.1.1
ujson==5.1.0
urllib3==1.26.9
uvloop==0.16.0
websockets==10.2
wrapt==1.14.0
